Backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.11.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
environ==1.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.1
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-secret-manager==2.24.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
mssql-django==1.6
mysqlclient==2.2.6
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.4.13
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
psycopg==3.2.9
psycopg-binary==3.2.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.0
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
redis==5.2.0
referencing==0.36.2
regex==2025.7.34
reportlab==4.2.5
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.7
sqlparse==0.5.2
stack-data==0.6.3
sympy==1.14.0
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.10
